move partition size related stuff to its own module
This avoids a circular import in another branch I am working on.
